2010-09-08  Rafael Espindola  <espindola@google.com>
	* script-sections.cc (Script_sections::add_memory_region): Convert
	field precision to int.
	* script.cc (script_set_section_region, script_set_section_region):
	Convert field precision to int.

+++ gold/script-sections.cc
@@ -2788,7 +2788,8 @@ Script_sections::add_memory_region(const char* name, size_t namelen,
     this->memory_regions_ = new Memory_regions();
   else if (this->find_memory_region(name, namelen))
     {
-      gold_error (_("region '%.*s' already defined"), namelen, name);
+      gold_error (_("region '%.*s' already defined"), static_cast<int>(namelen),
+                  name);
       // FIXME: Add a GOLD extension to allow multiple regions with the same
       // name.  This would amount to a single region covering disjoint blocks
       // of memory, which is useful for embedded devices.
